bot_app/requirements.txt
aiohttp==3.9.3
botbuilder-core==4.15.0
botbuilder-dialogs==4.15.0
botbuilder-integration-aiohttp==4.15.0
azure-storage-blob==12.20.0
botframework-connector==4.15.0
botframework-streaming==4.15.0
PyJWT==2.8.0
aiosignal==1.3.1
aiounittest==1.3.0
attrs==23.2.0
azure-core==1.30.2
babel==2.9.1
botbuilder-schema==4.15.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
datedelta==1.4
emoji==1.7.0
frozenlist==1.4.1
grapheme==0.6.0
idna==3.7
isodate==0.6.1
jsonpickle==1.4.2
msal==1.29.0
msrest==0.7.1
multidict==6.0.5
multipledispatch==1.0.0
oauthlib==3.2.2
pycparser==2.22
python-dotenv==0.19.0
pytz==2024.1
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
typing-extensions==4.12.2
ujson==5.10.0
urllib3==1.26.19
yarl==1.9.4
